+++ pkg/metrics/policy_report.go
@@ -17,7 +17,7 @@ func CreatePolicyReportMetricsCallback() report.PolicyReportCallback {
 	ruleGauge := promauto.NewGaugeVec(prometheus.GaugeOpts{
 		Name: "policy_report_result",
 		Help: "List of all PolicyReport Results",
-	}, []string{"namespace", "rule", "policy", "report", "kind", "name", "status"})
+	}, []string{"namespace", "rule", "policy", "report", "kind", "name", "status", "severity"})
 
 	prometheus.Register(policyGauge)
 	prometheus.Register(ruleGauge)
@@ -38,6 +38,7 @@ func CreatePolicyReportMetricsCallback() report.PolicyReportCallback {
 						res.Kind,
 						res.Name,
 						rule.Status,
+						rule.Severity,
 					).
 					Set(1)
 			}
@@ -54,6 +55,7 @@ func CreatePolicyReportMetricsCallback() report.PolicyReportCallback {
 					res.Kind,
 					res.Name,
 					rule.Status,
+					rule.Severity,
 				)
 			}
 
@@ -68,6 +70,7 @@ func CreatePolicyReportMetricsCallback() report.PolicyReportCallback {
 						res.Kind,
 						res.Name,
 						rule.Status,
+						rule.Severity,
 					).
 					Set(1)
 			}
@@ -89,6 +92,7 @@ func CreatePolicyReportMetricsCallback() report.PolicyReportCallback {
 					res.Kind,
 					res.Name,
 					rule.Status,
+					rule.Severity,
 				)
 			}
 		}